+++ fabric/src/main/java/com/unrealdinnerbone/jamd/data/BlockTagProvider.java
@@ -19,13 +19,13 @@ public class BlockTagProvider extends FabricTagProvider.BlockTagProvider {
     @Override
     protected void addTags(HolderLookup.Provider arg) {
         getOrCreateTagBuilder(BlockTags.MINEABLE_WITH_PICKAXE)
-                .add(JAMDRegistry.OVERWORLD.block().get())
-                .add(JAMDRegistry.NETHER.block().get())
-                .add(JAMDRegistry.END.block().get());
+                .add(JAMDRegistry.OVERWORLD.getBlock().get())
+                .add(JAMDRegistry.NETHER.getBlock().get())
+                .add(JAMDRegistry.END.getBlock().get());
         getOrCreateTagBuilder(BlockTags.NEEDS_DIAMOND_TOOL)
-                .add(JAMDRegistry.OVERWORLD.block().get())
-                .add(JAMDRegistry.NETHER.block().get())
-                .add(JAMDRegistry.END.block().get());
+                .add(JAMDRegistry.OVERWORLD.getBlock().get())
+                .add(JAMDRegistry.NETHER.getBlock().get())
+                .add(JAMDRegistry.END.getBlock().get());
 
     }
 }