+++ fabric/src/main/java/com/unrealdinnerbone/jamd/data/JAMDData.java
@@ -1,16 +1,22 @@
 package com.unrealdinnerbone.jamd.data;
 
+import com.unrealdinnerbone.jamd.JAMD;
 import com.unrealdinnerbone.jamd.JAMDRegistry;
 import net.fabricmc.fabric.api.datagen.v1.DataGeneratorEntrypoint;
 import net.fabricmc.fabric.api.datagen.v1.FabricDataGenerator;
+import net.minecraft.core.Holder;
 import net.minecraft.core.HolderGetter;
 import net.minecraft.core.RegistrySetBuilder;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.data.worldgen.BiomeDefaultFeatures;
 import net.minecraft.data.worldgen.BootstapContext;
 import net.minecraft.data.worldgen.biome.OverworldBiomes;
+import net.minecraft.data.worldgen.features.OreFeatures;
 import net.minecraft.data.worldgen.placement.NetherPlacements;
 import net.minecraft.data.worldgen.placement.OrePlacements;
+import net.minecraft.data.worldgen.placement.PlacementUtils;
+import net.minecraft.resources.ResourceKey;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.sounds.Musics;
 import net.minecraft.sounds.SoundEvents;
 import net.minecraft.tags.BlockTags;
@@ -19,7 +25,10 @@ import net.minecraft.world.level.biome.*;
 import net.minecraft.world.level.dimension.BuiltinDimensionTypes;
 import net.minecraft.world.level.dimension.DimensionType;
 import net.minecraft.world.level.levelgen.GenerationStep;
+import net.minecraft.world.level.levelgen.VerticalAnchor;
 import net.minecraft.world.level.levelgen.carver.ConfiguredWorldCarver;
+import net.minecraft.world.level.levelgen.feature.ConfiguredFeature;
+import net.minecraft.world.level.levelgen.placement.HeightRangePlacement;
 import net.minecraft.world.level.levelgen.placement.PlacedFeature;
 
 import java.util.OptionalLong;
@@ -60,7 +69,7 @@ public class JAMDData implements DataGeneratorEntrypoint {
             HolderGetter<PlacedFeature> placedFeatures = context.lookup(Registries.PLACED_FEATURE);
             HolderGetter<ConfiguredWorldCarver<?>> configuredWorldCarvers = context.lookup(Registries.CONFIGURED_CARVER);
             BiomeGenerationSettings.Builder builder = new BiomeGenerationSettings.Builder(placedFeatures, configuredWorldCarvers);
-            context.register(JAMDRegistry.Keys.END.biome(), new Biome.BiomeBuilder()
+            context.register(JAMDRegistry.END.getKey().biome(), new Biome.BiomeBuilder()
                     .temperature(0.5f)
                     .downfall(0.5f)
                     .hasPrecipitation(false)
@@ -77,7 +86,7 @@ public class JAMDData implements DataGeneratorEntrypoint {
         }
 
         static void bootstrapDimensionType(BootstapContext<DimensionType> context) {
-            context.register(JAMDRegistry.Keys.END.dimensionType(), new DimensionType(OptionalLong.of(6000),
+            context.register(JAMDRegistry.END.getKey().dimensionType(), new DimensionType(OptionalLong.of(6000),
                     false,
                     false,
                     false,
@@ -106,10 +115,10 @@ public class JAMDData implements DataGeneratorEntrypoint {
             BiomeDefaultFeatures.addExtraEmeralds(builder);
             builder.addFeature(GenerationStep.Decoration.UNDERGROUND_ORES, OrePlacements.ORE_COPPER_LARGE);
             builder.addFeature(GenerationStep.Decoration.UNDERGROUND_ORES, OrePlacements.ORE_TUFF);
-            context.register(JAMDRegistry.Keys.OVERWORLD.biome(), new Biome.BiomeBuilder()
+            context.register(JAMDRegistry.OVERWORLD.getKey().biome(), new Biome.BiomeBuilder()
                     .temperature(1)
                     .downfall(0.4f)
-                    .hasPrecipitation(true)
+                    .hasPrecipitation(false)
                     .temperatureAdjustment(Biome.TemperatureModifier.NONE)
                     .specialEffects(new BiomeSpecialEffects.Builder()
                             .skyColor(8103167)
@@ -123,7 +132,7 @@ public class JAMDData implements DataGeneratorEntrypoint {
         }
 
         static void bootstrapDimensionType(BootstapContext<DimensionType> context) {
-            context.register(JAMDRegistry.Keys.OVERWORLD.dimensionType(), new DimensionType(OptionalLong.of(6000),
+            context.register(JAMDRegistry.OVERWORLD.getKey().dimensionType(), new DimensionType(OptionalLong.of(6000),
                     true,
                     false,
                     false,
@@ -149,7 +158,7 @@ public class JAMDData implements DataGeneratorEntrypoint {
             HolderGetter<ConfiguredWorldCarver<?>> configuredWorldCarvers = context.lookup(Registries.CONFIGURED_CARVER);
             BiomeGenerationSettings.Builder builder = new BiomeGenerationSettings.Builder(placedFeatures, configuredWorldCarvers);
             BiomeDefaultFeatures.addNetherDefaultOres(builder);
-            context.register(JAMDRegistry.Keys.NETHER.biome(), new Biome.BiomeBuilder()
+            context.register(JAMDRegistry.NETHER.getKey().biome(), new Biome.BiomeBuilder()
                     .hasPrecipitation(false)
                     .temperature(2.0F)
                     .downfall(0.0F)
@@ -168,7 +177,7 @@ public class JAMDData implements DataGeneratorEntrypoint {
         }
 
         static void bootstrapDimensionType(BootstapContext<DimensionType> context) {
-            context.register(JAMDRegistry.Keys.NETHER.dimensionType(), new DimensionType(OptionalLong.of(18000),
+            context.register(JAMDRegistry.NETHER.getKey().dimensionType(), new DimensionType(OptionalLong.of(18000),
                     false,
                     false,
                     true,