+++ fabric/src/main/java/com/unrealdinnerbone/jamd/data/LangProvider.java
@@ -12,9 +12,9 @@ public class LangProvider extends FabricLanguageProvider {
 
     @Override
     public void generateTranslations(TranslationBuilder translationBuilder) {
-        translationBuilder.add(JAMDRegistry.OVERWORLD.block().get(), "Mining Portal");
-        translationBuilder.add(JAMDRegistry.NETHER.block().get(), "Nether Mining Portal");
-        translationBuilder.add(JAMDRegistry.END.block().get(), "End Mining Portal");
+        translationBuilder.add(JAMDRegistry.OVERWORLD.getBlock().get(), "Mining Portal");
+        translationBuilder.add(JAMDRegistry.NETHER.getBlock().get(), "Nether Mining Portal");
+        translationBuilder.add(JAMDRegistry.END.getBlock().get(), "End Mining Portal");
         translationBuilder.add("jamd.invalid.world", "Unable to find world '%s'");
         translationBuilder.add("jamd.invalid.pos", "Unable to find valid portal location");
         translationBuilder.add("biome.jamd.mining", "Mining");