+++ forge/src/main/java/com/unrealdinnerbone/jamd/JAMDForge.java
@@ -1,14 +1,33 @@
 package com.unrealdinnerbone.jamd;
 
-import com.unrealdinnerbone.trenzalore.api.config.ConfigManger;
+import com.unrealdinnerbone.jamd.api.FeatureTypeRegistry;
+import com.unrealdinnerbone.jamd.command.JamdCommand;
+import com.unrealdinnerbone.jamd.compact.MekenismOreCompact;
+import net.minecraftforge.common.MinecraftForge;
+import net.minecraftforge.event.RegisterCommandsEvent;
+import net.minecraftforge.event.server.ServerStartingEvent;
 import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 
 @Mod(JAMD.MOD_ID)
 public class JAMDForge {
-    
+
     public JAMDForge() {
         JAMD.init();
-        FMLJavaModLoadingContext.get().getModEventBus().addListener(JAMDDataForge::onData);
+        MinecraftForge.EVENT_BUS.addListener(this::onRegisterCommands);
+        MinecraftForge.EVENT_BUS.addListener(this::onStart);
+        registerCompact();
     }
+
+    public static void registerCompact() {
+        FeatureTypeRegistry.register("mekanism", MekenismOreCompact::new);
+    }
+
+    public void onRegisterCommands(RegisterCommandsEvent event) {
+        JamdCommand.register(event.getDispatcher());
+    }
+
+    public void onStart(ServerStartingEvent event) {
+        JAMD.onServerStart(event.getServer());
+    }
+
 }