+++ CHANGELOG.md
@@ -1,3 +1,8 @@
-### 3.2.2
+### 3.3.0
 
-- Blacklist Enlightened End mod
+- Rewrite mod to better support modded / custom ores
+- Added /jamd reload to reload the config
+- Added configs jamd/overworld.json, jamd/nether.json, jamd/end.json
+- Configs now completely controls the world generation
+- Added support for ore multipliers
+- Added Mekanism support